requirements.txt
streamlit==1.37.0
pandas==2.2.2
duckdb==1.0.0
pyarrow==17.0.0
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
tenacity==9.0.0
plotly==5.24.0

